move watch file and only bundle the sass we need
