+++ package.json
@@ -88,6 +88,7 @@
     "requireindex": "^1.1.0",
     "requirejs": "^2.3.2",
     "rimraf": "^2.5.4",
+    "sass-graph": "^2.1.2",
     "sass-lint": "^1.9.1",
     "semver": "^5.3.0",
     "sorted-object": "^1.0.0",
@@ -105,7 +106,6 @@
   },
   "scripts": {
     "postinstall": "./tools/task-runner/runner githooks --verbose",
-    "babel-watch": "BABEL_ENV=development babel ./static/src/javascripts --watch -d ./static/transpiled/javascripts >/dev/null 2>&1",
-    "watch": "./dev/start-dev-server.js"
+    "babel-watch": "BABEL_ENV=development babel ./static/src/javascripts --watch -d ./static/transpiled/javascripts >/dev/null 2>&1"
   }
 }
